Adding the fix for int

+++ spectrocloud/kubevirt/schema/virtualmachineinstance/domain_spec.go
@@ -113,10 +113,10 @@ func expandDisksToVM(disks []interface{}) []*models.V1VMDisk {
 			disk.Serial = v
 		}
 		if v, ok := in["boot_order"].(int); ok && v > 0 {
-			if v > math.MaxInt32 {
+			if uint64(v) > uint64(math.MaxInt32) {
 				disk.BootOrder = math.MaxInt32
 			} else {
-				disk.BootOrder = int32(v)
+				disk.BootOrder = int32(v) // #nosec G115 -- v > 0 and bounded by math.MaxInt32 above
 			}
 		}
 		if v, ok := in["disk_device"].([]interface{}); ok && len(v) > 0 {
